+++ visualization/app/codeCharta/state/store.service.ts
@@ -1,20 +1,8 @@
 import { IRootScopeService } from "angular"
 import { splitStateActions } from "./store/state.splitter"
-import { IsLoadingMapActions, setIsLoadingMap } from "./store/appSettings/isLoadingMap/isLoadingMap.actions"
-import { IsLoadingFileActions } from "./store/appSettings/isLoadingFile/isLoadingFile.actions"
 import { CCAction, State } from "../codeCharta.model"
-import { SearchPanelModeActions } from "./store/appSettings/searchPanelMode/searchPanelMode.actions"
-import { isActionOfType } from "../util/reduxHelper"
-import { SortingOrderAscendingActions } from "./store/appSettings/sortingOrderAscending/sortingOrderAscending.actions"
-import { SortingOptionActions } from "./store/dynamicSettings/sortingOption/sortingOption.actions"
-import { IsAttributeSideBarVisibleActions } from "./store/appSettings/isAttributeSideBarVisible/isAttributeSideBarVisible.actions"
-import { PanelSelectionActions } from "./store/appSettings/panelSelection/panelSelection.actions"
-import { PresentationModeActions } from "./store/appSettings/isPresentationMode/isPresentationMode.actions"
-import { ExperimentalFeaturesEnabledActions } from "./store/appSettings/enableExperimentalFeatures/experimentalFeaturesEnabled.actions"
 import { Store } from "./store/store"
-import { ScreenshotToClipboardEnabledActions } from "./store/appSettings/enableClipboard/screenshotToClipboardEnabled.actions"
-import { HoveredBuildingPathActions } from "./store/appStatus/hoveredBuildingPath/hoveredBuildingPath.actions"
-import { unfocusAllNodes } from "./store/dynamicSettings/focusedNodePath/focusedNodePath.actions"
+import { EffectsModule } from "./angular-redux/effects/effects.module"
 
 export interface StoreSubscriber {
 	onStoreChanged(actionType: string)
@@ -47,29 +35,6 @@ export class StoreService {
 	}
 
 	dispatch(action: CCAction, options: DispatchOptions = { silent: false }) {
-		if (
-			!(
-				isActionOfType(action.type, IsLoadingMapActions) ||
-				isActionOfType(action.type, IsLoadingFileActions) ||
-				isActionOfType(action.type, SortingOrderAscendingActions) ||
-				isActionOfType(action.type, SearchPanelModeActions) ||
-				isActionOfType(action.type, SortingOptionActions) ||
-				isActionOfType(action.type, IsAttributeSideBarVisibleActions) ||
-				isActionOfType(action.type, PanelSelectionActions) ||
-				isActionOfType(action.type, PresentationModeActions) ||
-				isActionOfType(action.type, ExperimentalFeaturesEnabledActions) ||
-				isActionOfType(action.type, ScreenshotToClipboardEnabledActions) ||
-				isActionOfType(action.type, HoveredBuildingPathActions) ||
-				options.silent
-			)
-		) {
-			this.dispatch(setIsLoadingMap(true))
-		}
-
-		if (isActionOfType(action.type, IsLoadingFileActions)) {
-			this.dispatch(unfocusAllNodes())
-		}
-
 		for (const atomicAction of splitStateActions(action)) {
 			this.originalDispatch(atomicAction)
 			if (!options.silent) {
@@ -77,6 +42,8 @@ export class StoreService {
 				this.notifyExtended(atomicAction.type, atomicAction.payload)
 			}
 		}
+
+		EffectsModule.actions$.next(action)
 	}
 
 	getState(): State {